refact!: Extract DefaultValue mixin from Value mixin and remove default values for value and default
