+++ src/Form/Mixin/Value.php
@@ -8,14 +8,9 @@ use ReflectionProperty;
 trait Value
 {
 	/**
-	 * Default value for the field, which will be used when a page/file/user is created
-	 */
-	protected mixed $default = null;
-
-	/**
 	 * The value of the field
 	 */
-	protected mixed $value = null;
+	protected mixed $value;
 
 	/**
 	 * @deprecated 5.0.0 Use `::toStoredValue()` instead to receive
@@ -31,18 +26,6 @@ trait Value
 		}
 
 		return $this->toStoredValue();
-	}
-
-	/**
-	 * Returns the default value of the field
-	 */
-	public function default(): mixed
-	{
-		if (is_string($this->default) === false) {
-			return $this->default;
-		}
-
-		return $this->model->toString($this->default);
 	}
 
 	/**
@@ -159,11 +142,6 @@ trait Value
 		return $this->hasValue();
 	}
 
-	protected function setDefault(mixed $default): void
-	{
-		$this->default = $default;
-	}
-
 	/**
 	 * Submits a new value for the field.
 	 * Fields can overwrite this method to provide custom
@@ -188,7 +166,7 @@ trait Value
 			return null;
 		}
 
-		return $this->value;
+		return $this->value ?? null;
 	}
 
 	/**
